Refactored to pull event handlers into renderUI.test.mjs where we test the UI click handlers separate from the todoList function. This is more separate and more unit level testing. Previously we had inegration tests that clicked a button and verified the todoList had changed.
